+++ mrmsg/data.go
@@ -1,7 +1,6 @@
 package mrmsg
 
 import (
-	"fmt"
 	"sort"
 	"strings"
 )
@@ -11,19 +10,29 @@ type (
 	Data map[string]any
 )
 
+// ValueString - возвращает значение в виде строки для указанного ключа.
+// Если ключ не найден, то возвращается пустая строка.
+func (d Data) ValueString(key string) string {
+	if val, ok := d[key]; ok {
+		return ToString(val)
+	}
+
+	return ""
+}
+
 // String - преобразовывает данные в строку.
 func (d Data) String() string {
-	var buf strings.Builder
-
-	buf.WriteByte('{')
-
-	// предварительная сортировка ключей т.к. map не гарантирует порядок
+	// предварительная сортировка ключей т.к. map не гарантирует их порядок
 	keys := make([]string, 0, len(d))
 	for k := range d {
 		keys = append(keys, k)
 	}
 
 	sort.Strings(keys)
+
+	var buf strings.Builder
+
+	buf.WriteByte('{')
 
 	firstItem := true
 	for _, key := range keys {
@@ -35,7 +44,7 @@ func (d Data) String() string {
 
 		buf.WriteString(key)
 		buf.Write([]byte{':', ' '})
-		buf.WriteString(fmt.Sprintf("%v", d[key]))
+		buf.WriteString(ToString(d[key]))
 	}
 
 	buf.WriteByte('}')